# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for قرار — decision making app
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Team {
  id: number;
  name: string;
  ownerId: string;
  inviteCode: string;
  createdAt: string;
  memberCount: number;
}

export interface TeamInput {
  /** @minLength 1 */
  name: string;
  memberName: string;
}

export interface JoinTeamInput {
  /** @minLength 1 */
  code: string;
  memberName: string;
}

export interface TeamMember {
  id: number;
  teamId: number;
  userId: string;
  name: string;
  role: string;
  joinedAt: string;
}

export interface Decision {
  id: number;
  question: string;
  createdAt: string;
  teamId?: number;
  status: string;
  decidedOptionId?: number;
  decidedAt?: string;
  createdByName?: string;
  deletedAt?: string | null;
}

export interface TeamDetail {
  id: number;
  name: string;
  ownerId: string;
  inviteCode: string;
  createdAt: string;
  members: TeamMember[];
  decisions: Decision[];
}

export interface AuditEntry {
  id: number;
  teamId?: number;
  decisionId?: number;
  userId: string;
  actorName: string;
  action: string;
  detail: string;
  createdAt: string;
}

export interface Comment {
  id: number;
  decisionId: number;
  optionId?: number;
  userId: string;
  authorName: string;
  body: string;
  createdAt: string;
}

export interface CommentInput {
  /** @minLength 1 */
  body: string;
  authorName: string;
  optionId?: number;
}

export interface DecisionInput {
  /** @minLength 1 */
  question: string;
  teamId?: number;
  createdByName?: string;
}

export interface DecisionUpdate {
  /** @minLength 1 */
  question?: string;
}

export interface DecideInput {
  optionId: number;
}

export interface Option {
  id: number;
  decisionId: number;
  label: string;
  sortOrder: number;
}

export interface OptionInput {
  /** @minLength 1 */
  label: string;
}

export interface OptionUpdate {
  /** @minLength 1 */
  label?: string;
}

export interface Criterion {
  id: number;
  decisionId: number;
  label: string;
  weight: number;
  sortOrder: number;
}

export interface CriterionInput {
  /** @minLength 1 */
  label: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  weight: number;
}

export interface CriterionUpdate {
  /** @minLength 1 */
  label?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  weight?: number;
}

export interface Rating {
  id: number;
  optionId: number;
  criterionId: number;
  userId: string;
  score: number;
}

export interface RatingInput {
  optionId: number;
  criterionId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  score: number;
}

export interface OptionScore {
  optionId: number;
  label: string;
  totalScore: number;
  maxScore: number;
  percentage: number;
  isWinner?: boolean;
}

export interface DecisionScores {
  scores: OptionScore[];
  voterCount: number;
  myVoteComplete: boolean;
}

export interface DecisionDetail {
  id: number;
  question: string;
  createdAt: string;
  teamId?: number;
  status: string;
  decidedOptionId?: number;
  decidedAt?: string;
  createdByName?: string;
  isCreator?: boolean;
  options: Option[];
  criteria: Criterion[];
  ratings: Rating[];
  myRatings: Rating[];
}