# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for قرار — decision making app
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List teams the authenticated user belongs to
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
        "401":
          description: Unauthorized
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a new team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /teams/join:
    post:
      operationId: joinTeam
      tags: [teams]
      summary: Join a team using an invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinTeamInput"
      responses:
        "200":
          description: Joined team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Invalid invite code
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get team detail with members and decisions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamDetail"
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /teams/{id}/audit:
    get:
      operationId: getTeamAudit
      tags: [audit]
      summary: Get the audit log for a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audit entries, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEntry"
        "401":
          description: Unauthorized
        "404":
          description: Not found
...
components:
  schemas:
...
    Team:
      type: object
      required: [id, name, ownerId, inviteCode, createdAt, memberCount]
      properties:
        id:
          type: integer
        name:
          type: string
        ownerId:
          type: string
        inviteCode:
          type: string
        createdAt:
          type: string
        memberCount:
          type: integer
    TeamInput:
      type: object
      required: [name, memberName]
      properties:
        name:
          type: string
          minLength: 1
        memberName:
          type: string
    JoinTeamInput:
      type: object
      required: [code, memberName]
      properties:
        code:
          type: string
          minLength: 1
        memberName:
          type: string
    TeamMember:
      type: object
      required: [id, teamId, userId, name, role, joinedAt]
      properties:
        id:
          type: integer
        teamId:
          type: integer
        userId:
          type: string
        name:
          type: string
        role:
          type: string
        joinedAt:
          type: string
    TeamDetail:
      type: object
      required: [id, name, ownerId, inviteCode, createdAt, members, decisions]
      properties:
        id:
          type: integer
        name:
          type: string
        ownerId:
          type: string
        inviteCode:
          type: string
        createdAt:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/TeamMember"
        decisions:
          type: array
          items:
            $ref: "#/components/schemas/Decision"
    AuditEntry:
      type: object
      required: [id, userId, actorName, action, detail, createdAt]
      properties:
        id:
          type: integer
        teamId:
          type: integer
        decisionId:
          type: integer
        userId:
          type: string
        actorName:
          type: string
        action:
          type: string
        detail:
          type: string
        createdAt:
          type: string
...
    Decision:
      type: object
      required: [id, question, createdAt, status]
      properties:
        id:
          type: integer
        question:
          type: string
        createdAt:
          type: string
        teamId:
          type: integer
        status:
          type: string
        decidedOptionId:
          type: integer
        decidedAt:
          type: string
        createdByName:
          type: string
        deletedAt:
          type: string
          nullable: true
...