# lib/api-zod/src/generated/types/decision.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for قرار — decision making app
 * OpenAPI spec version: 0.1.0
 */

export interface Decision {
  id: number;
  question: string;
  createdAt: string;
  teamId?: number;
  status: string;
  decidedOptionId?: number;
  decidedAt?: string;
  createdByName?: string;
  deletedAt?: string | null;
}